third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLineElement-svgdom-y1-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y1' property of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.y1.baseVal.value is 100
PASS lineElement.y1.baseVal.value is 10
PASS successfullyParsed is true

TEST COMPLETE
